requirements.txt.txt
python-telegram-bot==20.3
openai
yt-dlp
aiohttp
moviepy
ffmpeg-python
